+++ src/Events/MapGeneratedEvent.cpp
@@ -6,8 +6,8 @@ const EventType MapGeneratedEvent::eventType = EventType(&MapGeneratedEvent::eve
 MapGeneratedEvent::MapGeneratedEvent(){
 
 }
-MapGeneratedEvent::MapGeneratedEvent(float timeStamp){
-  this -> data = unique_ptr<MapGeneratedEventData>(new MapGeneratedEventData( timeStamp));
+MapGeneratedEvent::MapGeneratedEvent(float timeStamp,const vector<vector<int>>& distances,const vector<int>& entrances){
+  this -> data = unique_ptr<MapGeneratedEventData>(new MapGeneratedEventData( timeStamp,distances,entrances));
 }
 
 const EventType& MapGeneratedEvent::getEventType() const{